refactor: 도메인 용어 통일 test-result -> judgment result
